refactor: LadderRunner 클래스 생성
